Added inputchecks for date & time and also fixed some issues
Added inputchecks for date and time. Also fixed an issue where dates and times were being organized in the wrong order (for example december (12) would be before march (3) because decembers first digit (1) was before march (3). Now if a day, month or hour value has only 1 digit, a 0 is added before it so the days, months and hours are organized in the correct order