Added importance feature, added icons for each category, fixed an issue that crashed the app
Added importance feature where user can set an importance level for each task (0-3) and the importance is displayed in the list with stars.

Added icons for each category. Category icons are now displayed in the listview with the category textfield

Fixed an issue where the app would crash if the user didnt select a date or time when adding a task (trying to add a task with no values)